Code refactoring #49
