+++ server/src/index.ts
@@ -1,17 +1,16 @@
-import * as _ from 'lodash'
 import './config'
+import * as _ from 'lodash'
 import express from 'express'
 import bodyParser from 'body-parser'
 import cors from 'cors'
 import session from 'express-session'
 import FileStoreSession from 'session-file-store'
-// import proxy from 'http-proxy-middleware'
 import proxy from 'express-http-proxy'
 import passport from 'passport'
 import { Strategy as KeycloakStrategy } from 'passport-keycloak-oauth2-oidc'
 import { ApolloServer } from 'apollo-server-express'
 
-import { mergeSchemaList } from './utils/mergeSchemaList'
+import { mergeSchemaList } from '@utils'
 import { AppModule } from './modules'
 
 import { routes as addAuthRoutes } from './routes/auth'
