+++ server/src/routes/auth.ts
@@ -1,24 +1,11 @@
 import express from 'express'
-import getUser from '../units/getUser'
-import registerUser from '../units/registerUser'
-import '../config'
+import { getUser, registerUser } from '@units'
+import { uuid } from '@utils'
 import { logger } from '@shared'
 
 const keycloakRealm = process.env.KEYCLOAK_REALM
 const keycloakEndpoint = process.env.KEYCLOAK_ENDPOINT
 const keycloakPort = process.env.KEYCLOAK_PORT
-
-function uuid () {
-  let s = []
-  const hexDigits = '0123456789abcdef'
-  for (let i = 0; i < 36; i++) {
-    s[i] = hexDigits.substr(Math.floor(Math.random() * 0x10), 1)
-  }
-  s[14] = '4'
-  s[19] = hexDigits.substr((s[19] & 0x3) | 0x8, 1)
-  s[8] = s[13] = s[18] = s[23] = '-'
-  return s.join('')
-}
 
 export function routes (app: any, passport: any, session: any, keycloak: boolean) {
