+++ server/src/shared/Logger.ts
@@ -1,6 +1,4 @@
 import { createLogger, format, transports } from 'winston'
-
-// Import Functions
 const { File, Console } = transports
 
 // Init Logger
