+++ server/src/shared/index.ts
@@ -1,1 +1,2 @@
 export * from './Logger'
+// export * from './minioClient'
