+++ server/src/shared/minioClient.ts
@@ -1,0 +1,11 @@
+import { Client } from 'minio'
+
+const client = new Client({
+  endPoint: process.env.MINIO_ENDPOINT,
+  port: Number(process.env.MINIO_PORT),
+  accessKey: process.env.MINIO_ACCESS_KEY,
+  secretKey: process.env.MINIO_SECRET_KEY,
+  useSSL: false // Default is true.
+})
+
+export const minioClient = client
