+++ server/src/units/index.ts
@@ -1,0 +1,2 @@
+export * from './registerUser'
+export * from './getUser'
